Once again succumb to react-docgen-typescript's unreliability
