+++ package.json
@@ -22,6 +22,7 @@
     "test:unit": "jest",
     "test:e2e": "tsx scripts/testgen.ts && playwright test",
     "build": "tsx scripts/build.ts",
+    "build:docs": "pnpm -s docgen && cd docs && pnpm -s build",
     "docgen": "tsx scripts/docgen.ts",
     "api-report:update": "pnpm tsc && api-extractor run --local",
     "lint": "eslint . --ext .js,.jsx,.ts,.tsx",
