+++ src/display/Plot.tsx
@@ -1,4 +1,5 @@
-import { OfX, OfY } from "./Plot/Simple"
+import { OfX } from "./Plot/OfX"
+import { OfY } from "./Plot/OfY"
 import { Parametric } from "./Plot/Parametric"
 import { VectorField } from "./Plot/VectorField"
 import { Inequality } from "./Plot/Inequality"
@@ -12,7 +13,8 @@ export const Plot = {
 }
 
 import type { InequalityProps } from "./Plot/Inequality"
-import type { OfXProps, OfYProps } from "./Plot/Simple"
+import type { OfXProps } from "./Plot/OfX"
+import type { OfYProps } from "./Plot/OfY"
 import type { ParametricProps } from "./Plot/Parametric"
 import type { VectorFieldProps } from "./Plot/VectorField"
 export type { InequalityProps, OfXProps, OfYProps, ParametricProps, VectorFieldProps }
