+++ vercel.json
@@ -1,0 +1,4 @@
+{
+  "buildCommand": "pnpm -s docgen && cd docs && pnpm -s build",
+  "outputDirectory": "docs/build"
+}
